requirements.txt
# ==========================================================================
# Deep Search AI Assistant -- Python Dependencies
# ==========================================================================
# Install with:  pip install -r requirements.txt
#
# Notes:
#   - Versions are lower-bounded, not pinned, to avoid CUDA conflicts.
#   - faiss-cpu is used (not faiss-gpu) to avoid CUDA version coupling.
#   - sentence-transformers pulls in torch; let pip resolve the version.
#   - openvino is optional at this stage and listed separately below.
# ==========================================================================

# -- Core data handling ---------------------------------------------------
numpy>=1.24
Pillow>=9.0
opencv-python>=4.7
tqdm>=4.60
PyYAML>=6.0

# -- HuggingFace datasets (Arrow format loading) -------------------------
datasets>=2.14

# -- OCR ------------------------------------------------------------------
pytesseract>=0.3.10

# -- Embeddings -----------------------------------------------------------
sentence-transformers>=2.2

# -- Vector store ---------------------------------------------------------
faiss-cpu>=1.7.4

# -- Optional: OpenVINO (uncomment when ready for acceleration) -----------
# openvino>=2023.1
# openvino-dev>=2023.1
# openvino-genai>=2024.0
# optimum-intel[openvino]>=1.12

# -- Optional: PaddleOCR (uncomment when upgrading from Tesseract) --------
# paddleocr>=2.7
# paddlepaddle>=2.5
